+++ CHANGELOG.md
@@ -51,6 +51,10 @@ Malli is in well matured [alpha](README.md#alpha).
 ; => "olipa kerran"
 ```
 
+* `malli.dev.pretty/explain` for pretty-printing explanations
+
+<img src="https://github.com/metosin/malli/blob/master/docs/img/pretty-explain.png">
+
 ## 0.7.5 (2021-12-19)
 
 * [clj-kondo 2021.12.16+](https://github.com/clj-kondo/clj-kondo/blob/master/CHANGELOG.md#20211216) can load malli type configs automatically from new location (`.clj-kondo/metosin/malli-types/config.edn`)